+++ src/vs/editor/contrib/colorPicker/browser/standaloneColorPicker/standaloneColorPickerController.ts
@@ -1,0 +1,65 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IContextKey, IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
+import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
+import { ICodeEditor } from '../../../../browser/editorBrowser.js';
+import { IEditorContribution } from '../../../../common/editorCommon.js';
+import { EditorContextKeys } from '../../../../common/editorContextKeys.js';
+import { StandaloneColorPickerWidget } from './standaloneColorPickerWidget.js';
+import { Disposable } from '../../../../../base/common/lifecycle.js';
+import { EditorContributionInstantiation, registerEditorContribution } from '../../../../browser/editorExtensions.js';
+
+export class StandaloneColorPickerController extends Disposable implements IEditorContribution {
+
+	public static ID = 'editor.contrib.standaloneColorPickerController';
+	private _standaloneColorPickerWidget: StandaloneColorPickerWidget | null = null;
+	private _standaloneColorPickerVisible: IContextKey<boolean>;
+	private _standaloneColorPickerFocused: IContextKey<boolean>;
+
+	constructor(
+		private readonly _editor: ICodeEditor,
+		@IContextKeyService _contextKeyService: IContextKeyService,
+		@IInstantiationService private readonly _instantiationService: IInstantiationService,
+	) {
+		super();
+		this._standaloneColorPickerVisible = EditorContextKeys.standaloneColorPickerVisible.bindTo(_contextKeyService);
+		this._standaloneColorPickerFocused = EditorContextKeys.standaloneColorPickerFocused.bindTo(_contextKeyService);
+	}
+
+	public showOrFocus() {
+		if (!this._editor.hasModel()) {
+			return;
+		}
+		if (!this._standaloneColorPickerVisible.get()) {
+			this._standaloneColorPickerWidget = this._instantiationService.createInstance(
+				StandaloneColorPickerWidget,
+				this._editor,
+				this._standaloneColorPickerVisible,
+				this._standaloneColorPickerFocused
+			);
+		} else if (!this._standaloneColorPickerFocused.get()) {
+			this._standaloneColorPickerWidget?.focus();
+		}
+	}
+
+	public hide() {
+		this._standaloneColorPickerFocused.set(false);
+		this._standaloneColorPickerVisible.set(false);
+		this._standaloneColorPickerWidget?.hide();
+		this._editor.focus();
+	}
+
+	public insertColor() {
+		this._standaloneColorPickerWidget?.updateEditor();
+		this.hide();
+	}
+
+	public static get(editor: ICodeEditor) {
+		return editor.getContribution<StandaloneColorPickerController>(StandaloneColorPickerController.ID);
+	}
+}
+
+registerEditorContribution(StandaloneColorPickerController.ID, StandaloneColorPickerController, EditorContributionInstantiation.AfterFirstRender);
