+++ src/vs/editor/editor.all.ts
@@ -13,8 +13,8 @@ import './contrib/caretOperations/browser/transpose.js';
 import './contrib/clipboard/browser/clipboard.js';
 import './contrib/codeAction/browser/codeActionContributions.js';
 import './contrib/codelens/browser/codelensController.js';
-import './contrib/colorPicker/browser/colorContributions.js';
-import './contrib/colorPicker/browser/standaloneColorPickerActions.js';
+import './contrib/colorPicker/browser/hoverColorPicker/hoverColorPickerContribution.js';
+import './contrib/colorPicker/browser/standaloneColorPicker/standaloneColorPickerActions.js';
 import './contrib/comment/browser/comment.js';
 import './contrib/contextmenu/browser/contextmenu.js';
 import './contrib/cursorUndo/browser/cursorUndo.js';
